Open Keepass Database handling - work in progress
There will be a simplified mode where per default a database file will be automaticall loaded on application startup. This is for users which don't care about the particular database file on the phone but just want to store their passwords somewhere. And an extended mode where the user can choose between different database files which will be shown in a recently opened list view. That all should be configurable for expert users but hidden for the average user.